requirements.txt
numpy==1.26.2
openai==0.28.0
pandas==2.1.4
scikit_learn==1.3.2
torch==2.1.2
tqdm==4.66.1
transformers==4.36.2
weaviate_client==4.4b2
